Fix for Swamp Big Chest Logic
Modified the logic for the Big Chests further than my previous update as my previous change was insufficient.

This new code checks that the Big Key is properly accessible in the first 3/4 of Swamp Palace.  If it's not, then it makes sure you have Hookshot and that the Big Chest does not contain said hookshot.  As long as those two items are true you can get to the remaining chests and will be able to open the Big Chest.