Swamp Palace and Magic Bat Fixes
Removed the additional checks on the Big Chest in swamp palace and restored it to it's original logic from v7
Added a "CanAccessLateSwampPalace" function which does the following:

Checks that you have the hammer
Then it checks for the big key in any of the other chests in Swamp Palace that don't require the big key
If none of them have the big key THEN it checks if you have the hookshot.
If you have the hookshot it then makes sure that it's not in the big chest.

I associated "CanAccessLateSwampPalace(have)" to the requirements for the two chests in the flooded room, and the Hidden Waterfall door room.

I regenerated the seed with this being the ONLY modification to the code from v7 and the hookshot was still in the big chest but the Big Key moved to the Big Key chest.  I created 100 more seeds, all of which created without any Randomization Exceptions and the big key was all over Swamp Palace, including the three late-dungeon chests.  This feels like a far superio